made the 'gql_client' for GraphInterfaceBase init optional.
Manually specifying the gel client for the GraphInterface family of
classes is redundant in most cases, so I made it optional. When not
specified, we use the root_client's gql client (i.e. we inherit the GQL
client from the main Panther object.) We still want to be able to
manually specify the GQL client in some cases, since we use fake clients
for our unit tests.