refactor: base interface class (#8)
* added GraphInterfaceBase & reformatted sources.py

* renamed 'execute' to 'execute_gql'

In GraphInterfaceBase class, the function 'execute', which performs a
gql query and returns the results, would have a name conflict when the
'QueriesInterface' class is refactored to inherit from it. As such, we
rename 'execute' to 'execute_gql' to avoid the name collision.

* reformatted queries.py

* reformatted alerts.py

* reformatted cloud_accounts.py

* reformatted databases.py

* made the 'gql_client' for GraphInterfaceBase init optional.

Manually specifying the gel client for the GraphInterface family of
classes is redundant in most cases, so I made it optional. When not
specified, we use the root_client's gql client (i.e. we inherit the GQL
client from the main Panther object.) We still want to be able to
manually specify the GQL client in some cases, since we use fake clients
for our unit tests.

* reformatted users.py

* linting and formatting fixes

* removed module-level execute_gql

We can remove this function since it's been replaced by the class-level
version.