refactor: renamed panther.queries to panther.search (#19)
Since Panther released a REST API framework, the queries namespace we chose for querying the datalake now conflicts with the queries namespace for managing saved queries. To match as closely as possible to the vanilla Panther API, we opted to adjust our queries namespace to align with the REST API and rename our existing functionality to search.

This PR performs the migration of existing code from queries to search, and a future PR will add new functionality to the queries namespace.